Upgraded project settings for Xcode 5 / iOS 7 SDK
- Deferred creation of selectors to runtime and changed integer types to
  avoid compiler warnings.

+++ Source/CDRDefaultReporter.m
@@ -166,10 +166,10 @@
 
 - (void)printNestedFullTextForExample:(CDRExample *)example stateToken:(NSString *)token {
     static NSMutableArray *previousBranch = nil;
-    int previousBranchLength = previousBranch.count;
+    NSUInteger previousBranchLength = previousBranch.count;
 
     NSMutableArray *exampleBranch = [example fullTextInPieces];
-    int exampleBranchLength = exampleBranch.count;
+    NSUInteger exampleBranchLength = exampleBranch.count;
 
     BOOL onPreviousBranch = YES;
